Se implementa decorator

+++ src/juego/Caballero.java
@@ -39,7 +39,7 @@ public class Caballero extends Unidad {
 	}
 
 	@Override
-	public void fueAtacado(int daño) {
+	public void fueAtacado(double daño) {
 		this.salud -= (daño - this.defensa);
 	}
