+++ src/juego/Soldado.java
@@ -37,8 +37,9 @@ public class Soldado extends Unidad {
 	}
 
 	@Override
-	public void fueAtacado(int daño) {
+	public void fueAtacado(double daño) {
 		this.salud -= (daño - this.defensa);
 	}
+
 
 }
